+++ NEWS.md
@@ -1,10 +1,12 @@
 # derivmkts 0.2.2.1.0900
 
+* greeks `tidy` option renamed to `complete`
+
 * Added options to `greeks`: 
-  * `long` (if `tidy=TRUE`, return tidy long form output)
+  * `long` (if `complete=TRUE`, return long form output)
   * `initcaps`: capitalize "Delta", "Gamma", etc.
 
-* Changes:
+* Breaking change:
   * option value is now returned as "premium" rather than "price";
     the term "price" is ambiguous (the oil futures price is 100 but
     the value of the cotract --- the premium --- is 0)